update_calculate_time_Yike
I used the calling function. Give the computer information about the current time and waiting hours. Every 24 hours the time is cleared and reset. I need to add up the current time and waiting hours. The result will be divided by 24 and the remainder I obtain should be the final answer, which is YOUR ALARM WILL EXPIRE AT 14.